app_oil.py
from flask import Flask, request, abort

from linebot import (LineBotApi, WebhookHandler)
from linebot.exceptions import (InvalidSignatureError)
from linebot.models import *
# 在engine 的資料夾，找到 OilSearch
from engine.OilSearch import OilSearch

app = Flask(__name__)

# 設定你的Channel Access Token
line_bot_api = LineBotApi('vYB2Bf1brvu8VsdK06vU0uCiFyF1Xa8pg44fflfEqDvhpLooZF7H+/SvfYZDm0jXIpy+jmCjTQ1/RFPuvcZ9GWVKExdJnYItNgf+bspRK53t0AZCcIce9l0j9MY6NNtPxfE7d4i/OwVeebCglPJERwdB04t89/1O/w1cDnyilFU=')
# 設定你的Channel Secret
handler = WebhookHandler('0cfbfb198bb599f3945be3073913b429')

# 監聽所有來自 /callback 的 Post Request，我們不會動到
@app.route("/callback", methods=['POST'])
def callback():
    # get X-Line-Signature header value
    signature = request.headers['X-Line-Signature']
    # get request body as text
    body = request.get_data(as_text=True)
    app.logger.info("Request body: " + body)
    # handle webhook body
    try:
        handler.handle(body, signature)
    except InvalidSignatureError:
        abort(400)
    return 'OK'

#處理訊息
#當訊息種類為TextMessage時，從event中取出訊息內容，藉由TextSendMessage()包裝成符合格式的物件，並貼上message的標籤方便之後取用。
#接著透過LineBotApi物件中reply_message()方法，回傳相同的訊息內容。
#之後所有機器人判斷邏輯的編輯區
@handler.add(MessageEvent, message=TextMessage)
def handle_message(event):
    message = TextSendMessage(text=event.message.text)
    line_bot_api.reply_message(event.reply_token, message)

import os
if __name__ == "__main__":
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port)